# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/images/upload": {
      "post": {
        "operationId": "ImagesController_uploadImage",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload a product image",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images": {
      "get": {
        "operationId": "ImagesController_getUserImages",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user images",
        "tags": [
          "images"
        ]
      }
    },
    "/api/images/{id}": {
      "get": {
        "operationId": "ImagesController_getImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get image by ID",
        "tags": [
          "images"
        ]
      },
      "delete": {
        "operationId": "ImagesController_deleteImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an image and all its generations",
        "tags": [
          "images"
        ]
      }
    },
...
    "/api/payments/checkout": {
      "post": {
        "operationId": "PaymentsController_createCheckout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create Stripe checkout session for a one-time pack",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/subscribe": {
      "post": {
        "operationId": "PaymentsController_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create Stripe checkout session for a monthly plan",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/portal": {
      "post": {
        "operationId": "PaymentsController_portal",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stripe billing portal (cancel / change plan / update card)",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/subscription": {
      "get": {
        "operationId": "PaymentsController_subscription",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current subscription of the user (null when none)",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/history": {
      "get": {
        "operationId": "PaymentsController_getHistory",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get payment transaction history",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/webhook": {
      "post": {
        "operationId": "PaymentsController_handleWebhook",
        "parameters": [
          {
            "name": "stripe-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Stripe webhook endpoint (signature-verified, no JWT)",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/generation/styles": {
      "get": {
        "operationId": "GenerationController_getStyles",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Available styles and the default (starter) batch",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/generation/{imageId}/start": {
      "post": {
        "operationId": "GenerationController_startGeneration",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGenerationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start generating the selected styles (default: starter batch of 3)",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/generation/{imageId}/custom": {
      "post": {
        "operationId": "GenerationController_startCustomGeneration",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CustomGenerationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate a single custom image from user prompt + optional reference image",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/generation/retry/{id}": {
      "post": {
        "operationId": "GenerationController_retryGeneration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retry a failed generation (costs 1 credit)",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/generation/download/{id}": {
      "get": {
        "operationId": "GenerationController_downloadGeneration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download a generated image (proxy to avoid CORS)",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/generation/feedback/{id}": {
      "post": {
        "operationId": "GenerationController_submitFeedback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rate a finished graphic (thumbs up/down + reason)",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/generation/export/{id}": {
      "post": {
        "operationId": "GenerationController_exportGeneration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export a finished graphic in a marketplace format (ratio, size, promo badge)",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/generation/{imageId}/results": {
      "get": {
        "operationId": "GenerationController_getGenerations",
        "parameters": [
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get generation results for an image (polled by the UI)",
        "tags": [
          "generation"
        ]
      }
    },
    "/api/generation/result/{id}": {
      "get": {
        "operationId": "GenerationController_getGeneration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a single generation result",
        "tags": [
          "generation"
        ]
      }
    },
...
  },
  "info": {
    "title": "AllGrafika API",
    "description": "API for generating product graphics",
    "version": "1.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCheckoutDto": {
        "type": "object",
        "properties": {
          "packageId": {
            "type": "string",
            "example": "credits_15"
          },
          "acceptedWithdrawalWaiver": {
            "type": "boolean",
            "example": true,
            "description": "Zgoda na natychmiastowe udostępnienie kredytów i utratę prawa odstąpienia od umowy"
          }
        },
        "required": [
          "packageId",
          "acceptedWithdrawalWaiver"
        ]
      },
      "CreateSubscriptionDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "example": "sub_start"
          },
          "acceptedWithdrawalWaiver": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "planId",
          "acceptedWithdrawalWaiver"
        ]
      },
      "StartGenerationDto": {
        "type": "object",
        "properties": {
          "styles": {
            "description": "Identyfikatory stylów do wygenerowania. Domyślnie zestaw startowy (3 style).",
            "example": [
              "white-bg",
              "lifestyle-home",
              "dark-luxury"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "basePrompt": {
            "type": "string",
            "description": "Dodatkowe wskazówki stosowane do wszystkich stylów"
          }
        }
      },
      "CustomGenerationDto": {
        "type": "object",
        "properties": {
          "userPrompt": {
            "type": "string",
            "description": "Opis stylu / sceny lub żądanej modyfikacji"
          },
          "isRework": {
            "type": "boolean",
            "description": "true = przeróbka wcześniej wygenerowanej grafiki"
          }
        },
        "required": [
          "userPrompt"
        ]
      },
      "FeedbackDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "enum": [
              1,
              -1
            ],
            "description": "1 = kciuk w górę, -1 = kciuk w dół"
          },
          "reason": {
            "type": "string",
            "enum": [
              "product-changed",
              "artifacts",
              "wrong-style",
              "composition",
              "text-or-logo",
              "other"
            ]
          },
          "comment": {
            "type": "string",
            "description": "Opcjonalny komentarz"
          }
        },
        "required": [
          "rating"
        ]
      },
      "AdjustDto": {
        "type": "object",
        "properties": {
          "brightness": {
            "type": "number",
            "minimum": 0.5,
            "maximum": 1.5,
            "default": 1
          },
          "contrast": {
            "type": "number",
            "minimum": 0.5,
            "maximum": 1.5,
            "default": 1
          },
          "saturation": {
            "type": "number",
            "minimum": 0,
            "maximum": 2,
            "default": 1
          },
          "sharpen": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "CropDto": {
        "type": "object",
        "properties": {
          "left": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "top": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "width": {
            "type": "number",
            "minimum": 0.1,
            "maximum": 1
          },
          "height": {
            "type": "number",
            "minimum": 0.1,
            "maximum": 1
          }
        },
        "required": [
          "left",
          "top",
          "width",
          "height"
        ]
      },
      "ExportDto": {
        "type": "object",
        "properties": {
          "ratio": {
            "type": "string",
            "enum": [
              "1:1",
              "4:3",
              "3:4",
              "16:9"
            ],
            "default": "1:1"
          },
          "rotate": {
            "type": "number",
            "enum": [
              0,
              90,
              180,
              270
            ],
            "default": 0,
            "description": "Obrót zgodnie z ruchem wskazówek zegara"
          },
          "adjust": {
            "$ref": "#/components/schemas/AdjustDto"
          },
          "crop": {
            "description": "Kadr wybrany przez użytkownika; bez niego grafika jest dopasowana w całości (białe pasy)",
            "allOf": [
              {
                "$ref": "#/components/schemas/CropDto"
              }
            ]
          },
          "size": {
            "type": "number",
            "description": "Dłuższy bok w px (500–2560)",
            "default": 1600
          },
          "format": {
            "type": "string",
            "enum": [
              "jpeg",
              "png",
              "webp"
            ],
            "default": "jpeg"
          },
          "badgeText": {
            "type": "string",
            "description": "Tekst plakietki, np. \"-20%\" lub \"NOWOŚĆ\""
          },
          "badgeColor": {
            "type": "string",
            "enum": [
              "red",
              "orange",
              "green",
              "blue",
              "black"
            ]
          },
          "badgePosition": {
            "type": "string",
            "enum": [
              "top-left",
              "top-right",
              "bottom-left",
              "bottom-right"
            ]
          }
        }
      },
...
    }
  }
}